* Converted objDef_t::weight to int.
* To fix rounding errors (See bug #5404)

+++ src/game/chr_shared.h
@@ -99,7 +99,7 @@ typedef struct chrScoreMission_s {
 	int skillKills[SKILL_NUM_TYPES];	/**< Number of kills related to each skill. */
 
 	int heal;				/**< How many hitpoints has this soldier received trough healing in battlescape. */
-	float carriedWeight;	/**< The weight this soldier carried during combat */
+	int carriedWeight;	/**< The weight this soldier carried during combat */
 
 	inline chrScoreMission_s () {
 		init();
